Fix type inversion for shift and exp operators.

+++ test/libsolidity/syntaxTests/nameAndTypeResolution/shift_warn_literal_large_shift_amount.sol
@@ -1,6 +1,6 @@
 contract test {
     function f() pure public returns(uint) {
-        uint8 x = 100;
+        uint x = 100;
         return 10 << x;
     }
 }
